Error Fixes and Refactoring
Basic fixes for small items and some refactoring to be more in line with
single responsibility principle.
